Perceptual DM task
Filles to train RNNs in perceptual DM and code to visualize the results.  loop_call.py is the main file to train. It calls recurrent_main_to_loop.py  (to train) and generate_perceptual_dm.py (to generate the data set for the training). The same structure is for each considered task in the study. With loead_models_paper.py we can print the network parameters and testing samples. With print_status_2_inputs_paper.py we can print PCA and activity of individual units of the RNN.